@@ -1,28 +1,21 @@
 {
 	"name": "call-bind",
-	"version": "1.0.2",
+	"version": "1.0.7",
 	"description": "Robustly `.call.bind()` a function",
 	"main": "index.js",
 	"exports": {
-		".": [
-			{
-				"default": "./index.js"
-			},
-			"./index.js"
-		],
-		"./callBound": [
-			{
-				"default": "./callBound.js"
-			},
-			"./callBound.js"
-		],
+		".": "./index.js",
+		"./callBound": "./callBound.js",
 		"./package.json": "./package.json"
 	},
 	"scripts": {
-		"prepublish": "safe-publish-latest",
+		"prepack": "npmignore --auto --commentLines=auto",
+		"prepublish": "not-in-publish || npm run prepublishOnly",
+		"prepublishOnly": "safe-publish-latest",
 		"lint": "eslint --ext=.js,.mjs .",
+		"postlint": "evalmd README.md",
 		"pretest": "npm run lint",
-		"tests-only": "nyc tape 'test/*'",
+		"tests-only": "nyc tape 'test/**/*.js'",
 		"test": "npm run tests-only",
 		"posttest": "aud --production",
 		"version": "auto-changelog && git add CHANGELOG.md",
@@ -57,17 +50,31 @@
 	},
 	"homepage": "https://github.com/ljharb/call-bind#readme",
 	"devDependencies": {
-		"@ljharb/eslint-config": "^17.3.0",
-		"aud": "^1.1.3",
-		"auto-changelog": "^2.2.1",
-		"eslint": "^7.17.0",
+		"@ljharb/eslint-config": "^21.1.0",
+		"aud": "^2.0.4",
+		"auto-changelog": "^2.4.0",
+		"es-value-fixtures": "^1.4.2",
+		"eslint": "=8.8.0",
+		"evalmd": "^0.0.19",
+		"for-each": "^0.3.3",
+		"gopd": "^1.0.1",
+		"has-strict-mode": "^1.0.1",
+		"in-publish": "^2.0.1",
+		"npmignore": "^0.3.1",
 		"nyc": "^10.3.2",
-		"safe-publish-latest": "^1.1.4",
-		"tape": "^5.1.1"
+		"object-inspect": "^1.13.1",
+		"safe-publish-latest": "^2.0.0",
+		"tape": "^5.7.4"
 	},
 	"dependencies": {
-		"function-bind": "^1.1.1",
-		"get-intrinsic": "^1.0.2"
+		"es-define-property": "^1.0.0",
+		"es-errors": "^1.3.0",
+		"function-bind": "^1.1.2",
+		"get-intrinsic": "^1.2.4",
+		"set-function-length": "^1.2.1"
+	},
+	"testling": {
+		"files": "test/index.js"
 	},
 	"auto-changelog": {
 		"output": "CHANGELOG.md",
@@ -76,5 +83,13 @@
 		"commitLimit": false,
 		"backfillLimit": false,
 		"hideCredit": true
+	},
+	"publishConfig": {
+		"ignore": [
+			".github/workflows"
+		]
+	},
+	"engines": {
+		"node": ">= 0.4"
 	}
 }
