@@ -18,9 +18,15 @@
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/sequelize": "^4.28.15",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "eslint": "^8.56.0",
+    "eslint-config-standard-with-typescript": "^43.0.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
     "nodemon": "^2.0.22",
     "sequelize-cli": "^6.6.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.3"
   },
   "dependencies": {
     "dotenv": "^16.0.3",
