@@ -1,9 +1,9 @@
-import { Router } from "express";
-import Service from "../service/service.ts"
+import { Router } from 'express'
+import Service from '../service/service.ts'
 
-const router = Router();
-const service = new Service();
+const router = Router()
+const service = new Service()
 
-router.get('/check/:word', (req, res) => service.check(req, res));
+router.get('/check/:word', (req, res) => service.check(req, res))
 
-export { router };
+export { router }
