@@ -1,6 +1,6 @@
 import 'dotenv/config'
-import { App } from './app.ts';
+import { App } from './app.ts'
 
-const PORT = process.env.APP_PORT || 3001;
+const PORT = process.env.APP_PORT ?? 3001
 
-new App().start(PORT);
+new App().start(PORT)
